refactor: removing message deletion
This commit removes the message deletion when using a command as it
sends a message deletion embed every time someone uses a command.

Since Discord API does not let us know through audit logs or anything if
it is a bot who deleted a single message, the only way to avoid this log
spam is to completely remove the message deletion.